Refac: Apply Solid in Monitoria module
